Extract common & employer modules
Preparing for childminder side implementation.
Everything in common should be used in both sides.
Everything in employer should not.
